Refactor: deleted adapter package into presentation > file_viewer
