+++ storage-sample/src/main/res/layout/file_linear_adapter.xml
